requirements.txt
openai
mcp
jsonschema
python-dotenv
